u8 の使用を最小限に

+++ solver.hpp
@@ -34,7 +34,7 @@ public:
             node_count_ = 0;
             tt_hit_count_ = 0;
 
-            cerr << "Depth " << int(depth_max_) << ": ";
+            cerr << "Depth " << depth_max_ << ": ";
 
             u64 dur = bench([this,&pos]() { dfs(pos,0); });
             if(dur == 0) dur = 1;
@@ -53,14 +53,14 @@ public:
     }
 
 private:
-    void dfs(const Position& pos, u8 depth) {
+    void dfs(const Position& pos, int depth) {
         ++node_count_;
         auto pr = pos.result();
         if(pr == Board::SOLVED) {
             solution_all_.emplace_back(solution_);
             return;
         }
-        if(u8(depth+1) > depth_max_) return;
+        if(depth+1 > depth_max_) return;
         if(pr == Board::STUCK) return;
 
         u64 ar,up,down,left,right; tie(ar,up,down,left,right) = pos.moves();
@@ -77,7 +77,7 @@ private:
 
             Position pos2(pos);
             pos2.move_up(point);
-            dfs(pos2, u8(depth+1));
+            dfs(pos2, depth+1);
 
             solution_.pop_back();
             up &= ~point;
@@ -89,7 +89,7 @@ private:
 
             Position pos2(pos);
             pos2.move_down(point);
-            dfs(pos2, u8(depth+1));
+            dfs(pos2, depth+1);
 
             solution_.pop_back();
             down &= ~point;
@@ -101,7 +101,7 @@ private:
 
             Position pos2(pos);
             pos2.move_left(point);
-            dfs(pos2, u8(depth+1));
+            dfs(pos2, depth+1);
 
             solution_.pop_back();
             left &= ~point;
@@ -113,7 +113,7 @@ private:
 
             Position pos2(pos);
             pos2.move_right(point);
-            dfs(pos2, u8(depth+1));
+            dfs(pos2, depth+1);
 
             solution_.pop_back();
             right &= ~point;
@@ -121,7 +121,7 @@ private:
     }
 
     TranspositionTable tt_;
-    u8 depth_max_;
+    int depth_max_;
     u64 node_count_;
     u64 tt_hit_count_;
     vector<u64> solution_;
